Picoharp: Add -Wconversion to flags
Whoops: turns out we've been calling usleep(0) while waiting for the
picoharp to finish capturing.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/support/picoharp@107951 e099a375-04f9-0310-9d5f-a741eaff62e1

+++ picoharpApp/src/picoAsyn.c
@@ -246,7 +246,7 @@ static void picoThreadFunc(void *pvt)
     {
         /* acquire the data (usually 5s) */
         epicsMutexMustLock(pico->lock);
-        int pico_time = pico->data.time;
+        int pico_time = (int) pico->data.time;
         epicsMutexUnlock(pico->lock);
 
         /* clear error and measure */
